backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.10.0
email-validator>=2.0.0
alembic>=1.13.0
openpyxl>=3.1.2
reportlab>=4.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.0.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.9
supabase>=2.4.0
